pyproject.toml
[project]
name = "dailymotion"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "argon2-cffi>=25.1.0",
    "anyio>=4.13.0",
    "asyncpg>=0.31.0",
    "fastapi>=0.135.3",
    "pydantic[email]>=2.13.0",
    "pydantic-settings>=2.11.0",
    "redis>=6.4.0",
    "uvicorn>=0.44.0",
]

[dependency-groups]
tests = [
    "httpx>=0.28.0",
    "pytest>=9.0.3",
    "pytest-asyncio>=1.3.0",
    "testcontainers>=4.14.2",
]

[tool.pytest.ini_options]
pythonpath = ["."]
asyncio_mode = "auto"
asyncio_default_test_loop_scope = "session"
asyncio_default_fixture_loop_scope = "session"
